requirements.txt
oracledb==1.3.1
oracledb==1.3.2
pandas==2.0.3
pyodbc==4.0.39
python-dotenv==1.0.0
